pathologies: change how pre-selection works
pre-selection is used to mask out detectors in the low frequency
analysis before any crit-based cuts are applied. For dets that are not
pre-selected, their crit values are filled with zero.  In this commit
i removed this behavior by not filling the not pre-selected dets with
zero but compute their statistics, and only exclude the pre-selected
dets at the very end. This will allow me to treat all the crit on
equal footing and have pre-selection at my disposal as well.